% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_x_to_m.R
\name{cosine_x_to_m}
\alias{cosine_x_to_m}
\title{Cosine vector to matrix}
\usage{
cosine_x_to_m(x, m)
}
\arguments{
\item{x}{numeric vector}

\item{m}{numeric matrix}
}
\value{
vector of cosine similarities between x and m
}
\description{
Cosine vector to matrix
}
\examples{
vec <- c(1,2,3)
mat <- matrix(c(1,2,3,3,2,1,2,1,3),ncol=3)
cosine_x_to_m(vec,mat)

}
