% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article_article_similarities.R
\name{get_article_article_similarities}
\alias{get_article_article_similarities}
\title{Compute similarity between an abstract and all other abstracts}
\usage{
get_article_article_similarities(a_title, a_vectors = article_vectors,
  a_df = article_df)
}
\arguments{
\item{a_title}{charcter, the title of an abstract}

\item{a_vectors}{matrix, of semantic vectors for the abstracts (e.g., \code{article_vectors})}

\item{a_df}{dataframe, of abstract information (e.g., `article_df``)}
}
\value{
dataframe, of abstract information with Similiarities to target abstract appended
}
\description{
Compute similarity between an abstract and all other abstracts
}
\examples{

an_article <- article_df[1,]$title
sims <- get_article_article_similarities(an_article, article_vectors, article_df)
}
