% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mds_article_fits.R
\name{get_mds_article_fits}
\alias{get_mds_article_fits}
\title{Multidimensional scaling for article similarity}
\usage{
get_mds_article_fits(num_articles, num_clusters, year_range, input_df,
  a_vectors = article_vectors)
}
\arguments{
\item{num_articles}{integer, number of articles to select from entire dataframe}

\item{num_clusters}{integer, number of clusters to generate}

\item{year_range}{vector, c(1947,1948) containing beginning to end year}

\item{input_df}{dataframe an articles dataframe containing article info and Similarity info (e.g., like the one generated by \code{get_search_article_similarities()})}

\item{a_vectors}{matrix, containing the article vectors (e.g., \code{article_vectors})}
}
\value{
dataframe, a modified version of \code{input_df} containing MDS coordinates and cluster information useful for plotting
}
\description{
Multidimensional scaling for article similarity
}
\details{
columns are:
\itemize{
\item \code{formatted_column} for printing abstract info HTML
\item \code{title} abstract title
\item \code{wrap_title} useful for plotly hover labels
\item \code{year} abstract year
\item \code{index} abstract index
\item \code{Similarity} of search term to abstract
\item \code{X} x coordinate from MDS
\item \code{Y} y coodinate from MDS
\item \code{cluster} from k-means clustering
}
}
\examples{

search <- c("president")
article_dataframe<-get_search_article_similarities(search,query_type=1)
MDS <- get_mds_article_fits(10,2,c(1947,1948),article_dataframe)
}
