% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mds_author_fits.R
\name{get_mds_author_fits}
\alias{get_mds_author_fits}
\title{Multi-dimensional scaling for author similarity}
\usage{
get_mds_author_fits(num_authors, num_clusters, input_df,
  auth_vectors = AuthorVectors)
}
\arguments{
\item{num_authors}{integer, number of authors}

\item{num_clusters}{intger, number of clusters to find}

\item{input_df}{dataframe of abstract information (e.g., \code{article_df})}

\item{auth_vectors}{matrix of author vectors (e.g., \code{AuthorVectors})}
}
\value{
dataframe of abstract information with MDS coordinates and cluster information
}
\description{
Multi-dimensional scaling for author similarity
}
\examples{

author_sims <- get_author_similarities(author_list[1])
author_MDS <- get_mds_author_fits(15,3,author_sims,AuthorVectors)

}
