% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_terms.R
\name{get_search_terms}
\alias{get_search_terms}
\title{Get search terms in dictionary}
\usage{
get_search_terms(words, dictionary)
}
\arguments{
\item{words}{character a string with words}

\item{dictionary}{character vector containing all words in corpus}
}
\value{
character vector with words that are in the dictionary
}
\description{
Get search terms in dictionary
}
\details{
uses the breakdown function from LSAfun to get rid of special characters and convert to lowercase. The \code{dictionary_words} variable contains 100 sample words.
}
\examples{
search <- "foreward firetruck by the president"
corpus <- dictionary_words
get_search_terms(search,corpus)
}
