% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_vector.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Normalize a vector by max absolute value}
\usage{
normalize_vector(x)
}
\arguments{
\item{x}{numeric vector, a vector of numbers}
}
\value{
numeric vector, the vector divided by the largest absolute value in the vector
}
\description{
Normalize a vector by max absolute value
}
\examples{
a <- c(1,2,3)
normalize_vector(a)

}
