% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_article_vectors.R
\name{sl_article_vectors}
\alias{sl_article_vectors}
\title{Create article vectors from word vectors}
\usage{
sl_article_vectors(s_ids, w_matrix = word_vectors)
}
\arguments{
\item{s_ids}{list of integer vectors with ids for each word in article}

\item{w_matrix}{matrix of word vectors}
}
\value{
matrix of article vectors
}
\description{
Create article vectors from word vectors
}
\examples{

}
