% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_author_vectors.R
\name{sl_author_vectors}
\alias{sl_author_vectors}
\title{Create author vectors}
\usage{
sl_author_vectors(at_list, articles = article_df,
  a_vectors = article_vectors)
}
\arguments{
\item{at_list}{character vector of author names}

\item{articles}{dataframe with abstract info}

\item{a_vectors}{matrix of semantic article vectors}
}
\value{
matrix of author vectors (sum of abstracts)
}
\description{
Create author vectors
}
\examples{

}
