% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_beagle_vectors.R
\name{sl_beagle_vectors}
\alias{sl_beagle_vectors}
\title{Create BEAGLE vectors for words}
\usage{
sl_beagle_vectors(s_ids, dictionary = dictionary_words, riv = c(1024,
  30), verbose = 500)
}
\arguments{
\item{s_ids}{list of sentences as character vectors}

\item{dictionary}{character vector of all words in the dictionary}

\item{riv}{integer vector c(dim,sparsity) for number of dimensions and sparsity}

\item{verbose}{integer printing option to track progress report every n times}
}
\value{
matrix each row is the semantic vector for each word in dictionary
}
\description{
Create BEAGLE vectors for words
}
\examples{

}
