% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_clean.R
\name{sl_clean}
\alias{sl_clean}
\title{Clean words to lowercase etc.}
\usage{
sl_clean(words)
}
\arguments{
\item{words}{character}
}
\value{
character vector
}
\description{
Clean words to lowercase etc.
}
\examples{
sl_clean("Hello, this is a sentence I want to clean.")

}
