% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_corpus_dictionary.R
\name{sl_corpus_dictionary}
\alias{sl_corpus_dictionary}
\title{Get all unique words from corpus}
\usage{
sl_corpus_dictionary(words)
}
\arguments{
\item{words}{character vector}
}
\value{
character vector of unique words
}
\description{
uses the breakdown function from LSAfun, which converts to lowercase and removes most special characters. After breakdown, all unique words are returned, forming the dictionary for the corpus.
}
\examples{

sl_corpus_dictionary("The the three","more stuff")
}
