% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_create_riv.R
\name{sl_create_riv}
\alias{sl_create_riv}
\title{Create random index vectors}
\usage{
sl_create_riv(dim, inds, sparsity)
}
\arguments{
\item{dim}{integer, number of dimensions (vector length)}

\item{inds}{integer, number of rows}

\item{sparsity}{even integer, number of elements in vector}
}
\value{
matrix, each row is an item with dim dimensions
}
\description{
Create random index vectors
}
\examples{

sl_create_riv(10,5,1)
sl_create(10,5,4)
}
