% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_semantic_space.R
\name{sl_semantic_space}
\alias{sl_semantic_space}
\title{Create semantic space from article dataframe}
\usage{
sl_semantic_space(abstract_df, col_names, riv_param = c(1024, 30))
}
\arguments{
\item{abstract_df}{dataframe containing abstract info}

\item{col_names}{names of columns containing text (e.g., title, abstract)}

\item{riv_param}{integer vector (dims,sparsity)}
}
\value{
list containing
\itemize{
\item \code{dictionary_words} character vector of unique words in corpus
\item \code{word_vectors} matrix of semantic vectors for each word
\item \code{article_vectors} matrix of semantic vectors for each article
\item \code{author_list} character vector of authors
\item \code{author_vectors} matrix of semantic vectors for each author
}
}
\description{
Create semantic space from article dataframe
}
\examples{

}
