% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_word_ids_sentences.R
\name{sl_word_ids_sentences}
\alias{sl_word_ids_sentences}
\title{Get indexes for each word in sentence}
\usage{
sl_word_ids_sentences(sentences, words)
}
\arguments{
\item{sentences}{list of character vectors for each sentence}

\item{words}{dictionary of words in corpus}
}
\value{
list of index vectors for each word in sentence
}
\description{
Get indexes for each word in sentence
}
\examples{

my_words <- c("hello","this","is")
my_sentences <-list(c("hello","is"),c("this","is"))
sl_word_ids_sentences(my_sentences,my_words)

}
