% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_string.R
\name{wrap_string}
\alias{wrap_string}
\title{Wrap a string}
\usage{
wrap_string(x, n_chars, sep = " <br> ")
}
\arguments{
\item{x}{character, a string with words in it}

\item{n_chars}{numeric, number of characters before split}

\item{sep}{charcter, a character string with the separator, default is " <br> " for html}
}
\value{
character, a string, with the text wrapped (separate) by the \code{sep} delimiter
}
\description{
Wrap a string
}
\examples{
a <- "a really long string that you want to wrap"
wrap_string(a,n_chars = 20, sep = )
}
